Fixing details in csv action title
