+++ src/views/booking/steps/TableAttendees.js
@@ -31,6 +31,7 @@ import {
   UncontrolledButtonDropdown
 } from 'reactstrap';
 import ExportToExcel from '../../../components/ExportToExcel';
+import ExportToCsv from '../../../components/ExportToCsv';
 import { BOOKING_CLOSED_STATUS } from '../../../utility/Constants';
 
 // ** Bootstrap Checkbox Component
@@ -268,60 +269,6 @@ const DataTableAttendees = ({
     />
   );
 
-  // ** Converts table to CSV
-  function convertArrayOfObjectsToCSV (array) {
-    let result;
-    const columnDelimiter = ';';
-    const lineDelimiter = '\n';
-    const dynamicLabels = array.length > 0 && array[0].additionalFields && array[0].additionalFields.map((item) => item.name);
-    const keys = `name;email;phone;addressLine1;addressLine2;city;state;zip;country`;
-    const arraykeys = keys.split(';');
-    result = '';
-    result += arraykeys.join(columnDelimiter);
-    result += dynamicLabels && dynamicLabels.length > 0 ? `;${dynamicLabels.join(columnDelimiter)};bookingId` : `;bookingId`;
-    result += ``;
-    result += lineDelimiter;
-
-    array.forEach((item) => {
-      let ctr = 0;
-      arraykeys.forEach((key) => {
-        if (ctr > 0) result += columnDelimiter;
-        result += (item[key] && item[key].replace('#', '')) || '';
-        ctr++;
-      });
-      if (item.additionalFields && item.additionalFields.length > 0) {
-        result += columnDelimiter;
-        item.additionalFields.map((field) => {
-          result += field.value;
-          result += columnDelimiter;
-        });
-      } else {
-        result += columnDelimiter;
-      }
-      result += item['bookingId'];
-      result += lineDelimiter;
-    });
-    return result;
-  }
-  // ** Downloads CSV
-  function downloadCSV (array) {
-    const link = document.createElement('a');
-    let csv = convertArrayOfObjectsToCSV(array);
-    if (csv === null) return;
-
-    const filename = `${customer && customer.name}${customer && customer.company ? ', ' : ''}${
-      customer && customer.company ? customer.company : ''
-    }-${moment().format('LL')}-${teamClassInfo.title}.csv`;
-
-    if (!csv.match(/^data:text\/csv/i)) {
-      csv = `data:text/csv;charset=utf-8,${csv}`;
-    }
-
-    link.setAttribute('href', encodeURI(csv));
-    link.setAttribute('download', filename);
-    link.click();
-  }
-
   return (
     <Fragment>
       <Card>
@@ -361,11 +308,12 @@ const DataTableAttendees = ({
                     </DropdownItem>
                     {booking && booking.status !== BOOKING_CLOSED_STATUS && (
                       <DropdownItem className="w-100" onClick={handleModalUpload}>
-                        <Grid size={15} />
-                        <span className="align-middle ml-50">
-                          Upload data<br></br>
+                        <>
+                          <h6>
+                            <Grid size={15} /> Upload data
+                          </h6>
                           <small>Excel file with your attendees</small>
-                        </span>
+                        </>
                       </DropdownItem>
                     )}
                     <DropdownItem className="align-middle w-100">
@@ -383,12 +331,20 @@ const DataTableAttendees = ({
                         smallText={<h6 className="small m-0 p-0">Download excel file with attendees</h6>}
                       />
                     </DropdownItem>
-                    <DropdownItem onClick={() => downloadCSV(attendees)} className="align-middle w-100">
-                      <File size={13} />
-                      <span className="mb-1">CSV File</span>
-                      <small>
-                        <h6 className="small">Download excel file with attendees</h6>
-                      </small>
+                    <DropdownItem className="align-middle w-100">
+                      <ExportToCsv
+                        array={attendees}
+                        name={`${customer && customer.name}${customer && customer.company ? ', ' : ''}${
+                          customer && customer.company ? customer.company : ''
+                        }-${moment().format('LL')}-${teamClassInfo.title}.csv`}
+                        title={
+                          <h6>
+                            <FileText size={13} />
+                            {'   Csv File'}
+                          </h6>
+                        }
+                        smallText={<h6 className="small m-0 p-0">Download csv file with attendees</h6>}
+                      />
                     </DropdownItem>
                   </DropdownMenu>
                 </UncontrolledButtonDropdown>
